indices.simulate_template
