+++ specification/indices/simulate_template/IndicesSimulateTemplateResponse.ts
@@ -18,13 +18,14 @@
  */
 
 import { Alias } from '@indices/_types/Alias'
+import { IndexSettings } from '@indices/_types/IndexSettings'
 import { Dictionary } from '@spec_utils/Dictionary'
-import { UserDefinedValue } from '@spec_utils/UserDefinedValue'
 import { IndexName, Name } from '@_types/common'
 import { TypeMapping } from '@_types/mapping/TypeMapping'
 
 export class Response {
   body: {
+    overlapping?: Overlapping[]
     template: Template
   }
 }
@@ -32,8 +33,7 @@ export class Response {
 export class Template {
   aliases: Dictionary<IndexName, Alias>
   mappings: TypeMapping
-  settings: Dictionary<string, UserDefinedValue>
-  overlapping: Overlapping[]
+  settings: IndexSettings
 }
 
 export class Overlapping {
